requirements.txt
acme==1.1.0
Automat==0.8.0
blinker==1.4
boto3==1.20.41
botocore==1.23.41
certbot==0.40.0
cffi==1.15.1
Click==7.0
click-odoo==1.5.0
click-odoo-contrib==1.15.1
colorama==0.4.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
distro==1.8.0
future==0.18.2
importlib-resources==5.4.0
importmagic3==0.2.0
jedi==0.18.1
jmespath==0.10.0
josepy==1.2.0
log-colorizer==2.0.0
parsedatetime==2.4
parso==0.8.3
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-magic==0.4.27
s3transfer==0.5.0
spark-parser==1.8.9
uncompyle6==3.8.0
wdb==3.3.0
xdis==6.0.4
zipp==3.7.0
zope.component==4.3.0
zope.event==4.4
zope.interface==4.7.1
